Add files via upload
If you're like many people, you probably download a lot of files from the internet. These files can be anything from pictures to videos to documents, and they can quickly pile up in your computer's downloads folder. When this happens, it can be difficult and time-consuming to find the files you need.

To solve this problem, I created a script that automates the process of organizing files based on their file extension. The script uses a dictionary that maps file extensions to their corresponding folders, and then it loops through all the files in the downloads folder to find files that match these extensions. Once it finds a match, it moves the file to the appropriate folder.

With this script, I no longer have to spend hours organizing my downloads folder. Now, everything is automatically sorted and easy to find. Plus, I can add new file types and folders to the dictionary as needed, making the script even more useful and customizable.